organize files and add image to header
