+++ pokedex/src/App.js
@@ -1,11 +1,8 @@
-
-import Main from "./pages/components/Main";
 import { Router } from "./routes/Routes";
 
 function App() {
   return (
     <>
-    {/* <Main /> */}
     <Router/>
     </>
   );
