+++ pokedex/src/components/pokeinfo/styled.js
@@ -1,0 +1,22 @@
+import styled from "styled-components"
+
+export const Abilits = styled.div `
+width: 30%;
+margin: auto;
+display: flex;
+justify-content: space-around;
+align-items: center;
+margin-top: 1rem;
+`
+
+export const AbilitsGroup = styled.div `
+background-color: #b74555;
+color: white;
+padding: 0.5rem;
+font-size: 15px;
+border-radius: 8px;
+`
+
+export const BaseStat = styled.div `
+margin-top: 2rem
+`
