+++ pokedex/src/pages/detailspage/DetailsPage.jsx
@@ -1,5 +1,4 @@
 import React, { useEffect, useState } from "react";
-import styled from "styled-components";
 import { useParams } from "react-router-dom";
 import axios from "axios";
 import { Images } from "./styled";
