+++ pokedex/src/pages/homepage/style.css
@@ -10,7 +10,7 @@ body{
 .container{
     width: 100%;
     margin: auto;
-    padding-top: 100px;
+    padding-top: 40px;
     display: flex;
     background-color: #A83A3A;
     justify-content: center;
