+++ pokedex/src/pages/initialpage/InitialPage.js
@@ -2,15 +2,12 @@ import logo from "../../images/logo-pokemon.png";
 import { useNavigate } from 'react-router-dom';
 import { BackgroundPage, Img, ButtonHome } from "./styled";
 
-
-
 export default function InitialPage() {
     const navigate = useNavigate()   
 
     const goToHome = () => {
         navigate("/home")
     }
-
 
   return (
     <BackgroundPage className="App">
