+++ pokedex/src/routes/Routes.js
@@ -1,6 +1,6 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Main from "../pages/components/Main";
 import DetailsPage from "../pages/detailspage/DetailsPage";
+import HomePage from "../pages/homepage/HomePage";
 import InitialPage from "../pages/initialpage/InitialPage";
 
 export const Router = () => {
@@ -8,7 +8,7 @@ export const Router = () => {
     <BrowserRouter>
       <Routes>
         <Route index element={<InitialPage />} />
-        <Route path="/home" element={<Main />} />
+        <Route path="/home" element={<HomePage />} />
         <Route path="/details/:pokemonId" element={<DetailsPage />} />
       </Routes>
     </BrowserRouter>
